Bug fixes
Students can now see the quizzes which have been assigned to their group.

Quiz applet now takes an assignment id instead of a quiz id.  This means that a student who is assigned the same quiz twice will no longer have issues taking both.

Fixed an SSR hydration issue with macOS.

You can no longer leave the quiz page while your answers are being saved.

+++ src/graphql/resolvers/session.ts
@@ -7,8 +7,8 @@ import { ProtectQuery } from '../resolvers';
 export const Session = {
     typeDefs: gql`
         type QuizGrade {
-            grade: Int,
-            graded: Int,
+            grade: Float,
+            graded: Float,
         }
 
         extend type Query {
